moving functions to where they're actually used instead of having them at global scope
