+++ src/types/shortlist-it-state-manager.ts
@@ -1,0 +1,6 @@
+import { ShortlistItState } from "./shortlist-it-state";
+
+export type ShortlistItStateManager = {
+    readonly state: ShortlistItState,
+    setState: React.Dispatch<React.SetStateAction<ShortlistItState>>;
+};
